Implement IAdapter interface and enhance documentation
- Updated AddDivisionAdapter to implement IAdapter with Errors property and ExecuteAsync method.
- Enhanced AddDivisionPresenter with XML documentation and improved error handling.
- Modified IAddDivisionView to inherit from TournamentManager.IView and added method documentation.
- Restructured DivisionViewModel to include XML documentation and a constructor for initialization.
- Updated IRetrieveDivisionsView with method documentation for clarity.
- Modified RetrieveDivisionsAdapter to implement IAdapter and added Error property.
- Enhanced RetrieveDivisionsPresenter with XML documentation for ExecuteAsync and AddDivisionAsync methods.
