--- source/main/Divisions/Add/AddDivisionPresenter.cs
@@ -1,71 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace BowlingMegabucks.TournamentManager.Divisions.Add;
-internal class Presenter
-{
-    private readonly Lazy<Retrieve.IAdapter> _retrieveDivisionsAdapter;
-    private Retrieve.IAdapter RetrieveDivisionsAdapter => _retrieveDivisionsAdapter.Value;
-
-    private readonly Lazy<IAdapter> _addDivisionAdapter;
-    private IAdapter AddDivisionAdapter => _addDivisionAdapter.Value;
-
-    private readonly IView _view;
-
-    public Presenter(IView view, IServiceProvider services)
-    {
-        _view = view;
-
-        _retrieveDivisionsAdapter = new Lazy<Retrieve.IAdapter>(services.GetRequiredService<Retrieve.IAdapter>);
-        _addDivisionAdapter = new Lazy<IAdapter>(services.GetRequiredService<IAdapter>);
-    }
-
-    /// <summary>
-    /// Unit Test Constructor
-    /// </summary>
-    /// <param name="mockView"></param>
-    /// <param name="mockRetrieveDivisionsAdapter"></param>
-    /// <param name="mockAddDivisionAdapter"></param>
-    internal Presenter(IView mockView, Retrieve.IAdapter mockRetrieveDivisionsAdapter, IAdapter mockAddDivisionAdapter)
-    {
-        _view = mockView;
-        _retrieveDivisionsAdapter = new Lazy<Retrieve.IAdapter>(() => mockRetrieveDivisionsAdapter);
-        _addDivisionAdapter = new Lazy<IAdapter>(() => mockAddDivisionAdapter);
-    }
-
-    public async Task GetNextDivisionNumberAsync(CancellationToken cancellationToken)
-    {
-        var divisions = await RetrieveDivisionsAdapter.ExecuteAsync(_view.Division.TournamentId, cancellationToken).ConfigureAwait(true);
-
-        if (RetrieveDivisionsAdapter.Error != null)
-        {
-            _view.DisplayErrors([RetrieveDivisionsAdapter.Error.Message]);
-        }
-        else
-        {
-            _view.Division.Number = (short)(divisions.Count() + 1);
-        }
-    }
-
-    public async Task ExecuteAsync(CancellationToken cancellationToken)
-    {
-        if (!_view.IsValid())
-        {
-            _view.KeepOpen();
-            return;
-        }
-
-        var id = await AddDivisionAdapter.ExecuteAsync(_view.Division, cancellationToken).ConfigureAwait(true);
-
-        if (AddDivisionAdapter.Errors.Any())
-        {
-            _view.KeepOpen();
-            _view.DisplayErrors(AddDivisionAdapter.Errors.Select(e => e.Message));
-        }
-        else
-        {
-            _view.DisplayMessage($"{_view.Division.DivisionName} division added.");
-            _view.Division.Id = id!.Value;
-            _view.Close();
-        }
-    }
-}
