--- source/main/Divisions/Add/IAddDivisionView.cs
@@ -1,7 +1,0 @@
-﻿namespace BowlingMegabucks.TournamentManager.Divisions.Add;
-internal interface IView : BowlingMegabucks.TournamentManager.IView
-{
-    void DisplayErrors(IEnumerable<string> errorMessages);
-
-    IViewModel Division { get; }
-}
