+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Divisions/Retrieve/RetrieveDivisionsAdapter.cs
@@ -1,0 +1,27 @@
+﻿namespace BowlingMegabucks.TournamentManager.Divisions.Retrieve;
+internal class Adapter : IAdapter
+{
+    private readonly IBusinessLogic _businessLogic;
+
+    public Adapter(IBusinessLogic businessLogic)
+    {
+        _businessLogic = businessLogic;
+    }
+
+    public Models.ErrorDetail? Error
+        => _businessLogic.ErrorDetail;
+
+    public async Task<IEnumerable<IViewModel>> ExecuteAsync(TournamentId tournamentId, CancellationToken cancellationToken)
+    {
+        var divisions = await _businessLogic.ExecuteAsync(tournamentId, cancellationToken).ConfigureAwait(false);
+
+        return divisions.Select(division => new ViewModel(division));
+    }
+}
+
+internal interface IAdapter
+{
+    Models.ErrorDetail? Error { get; }
+
+    Task<IEnumerable<IViewModel>> ExecuteAsync(TournamentId tournamentId, CancellationToken cancellationToken);
+}
